Add types and update parse return type
The parse method now returns JavaScript objects instead of JSON strings.
With this, are some TypeScript definitions of the output. These types
are not guaranteed to be correct, so an additional integration test
should be performed to parse a maximally representative Tangibl AST and
ensure that the types correlate with the output.